requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cymem==2.0.5
Cython==0.29.21
decorator==4.4.2
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
gensim==3.8.1
gitdb==4.0.7
GitPython==3.1.14
grpcio==1.32.0
h5py==2.10.0
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.7.1
pathy==0.5.2
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prometheus-client==0.10.1
prompt-toolkit==3.0.8
protobuf==3.15.8
pyarrow==4.0.0
pycparser==2.20
pydantic==1.7.3
pydeck==0.6.2
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==228
pywinpty==1.0.0
pyzmq==19.0.2
regex==2021.4.4
requests==2.25.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
smart-open==3.0.0
smmap==4.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
streamlit==0.80.0
terminado==0.9.4
testpath==0.4.4
textblob==0.15.3
thinc==8.0.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typer==0.3.2
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
wasabi==0.8.2
watchdog==2.0.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
